+++ sdk/resourcemanager/datafactory/armdatafactory/models_serde.go
@@ -10211,6 +10211,57 @@ func (a *AzureTableStorageLinkedService) UnmarshalJSON(data []byte) error {
 	return nil
 }
 
+// MarshalJSON implements the json.Marshaller interface for type AzureTableStorageLinkedServiceTypeProperties.
+func (a AzureTableStorageLinkedServiceTypeProperties) MarshalJSON() ([]byte, error) {
+	objectMap := make(map[string]any)
+	populate(objectMap, "accountKey", a.AccountKey)
+	populateAny(objectMap, "connectionString", a.ConnectionString)
+	populate(objectMap, "credential", a.Credential)
+	populate(objectMap, "encryptedCredential", a.EncryptedCredential)
+	populate(objectMap, "sasToken", a.SasToken)
+	populateAny(objectMap, "sasUri", a.SasURI)
+	populateAny(objectMap, "serviceEndpoint", a.ServiceEndpoint)
+	return json.Marshal(objectMap)
+}
+
+// UnmarshalJSON implements the json.Unmarshaller interface for type AzureTableStorageLinkedServiceTypeProperties.
+func (a *AzureTableStorageLinkedServiceTypeProperties) UnmarshalJSON(data []byte) error {
+	var rawMsg map[string]json.RawMessage
+	if err := json.Unmarshal(data, &rawMsg); err != nil {
+		return fmt.Errorf("unmarshalling type %T: %v", a, err)
+	}
+	for key, val := range rawMsg {
+		var err error
+		switch key {
+		case "accountKey":
+			err = unpopulate(val, "AccountKey", &a.AccountKey)
+			delete(rawMsg, key)
+		case "connectionString":
+			err = unpopulate(val, "ConnectionString", &a.ConnectionString)
+			delete(rawMsg, key)
+		case "credential":
+			err = unpopulate(val, "Credential", &a.Credential)
+			delete(rawMsg, key)
+		case "encryptedCredential":
+			err = unpopulate(val, "EncryptedCredential", &a.EncryptedCredential)
+			delete(rawMsg, key)
+		case "sasToken":
+			err = unpopulate(val, "SasToken", &a.SasToken)
+			delete(rawMsg, key)
+		case "sasUri":
+			err = unpopulate(val, "SasURI", &a.SasURI)
+			delete(rawMsg, key)
+		case "serviceEndpoint":
+			err = unpopulate(val, "ServiceEndpoint", &a.ServiceEndpoint)
+			delete(rawMsg, key)
+		}
+		if err != nil {
+			return fmt.Errorf("unmarshalling type %T: %v", a, err)
+		}
+	}
+	return nil
+}
+
 // MarshalJSON implements the json.Marshaller interface for type BigDataPoolParametrizationReference.
 func (b BigDataPoolParametrizationReference) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]any)
@@ -12266,6 +12317,41 @@ func (c *ConnectionStateProperties) UnmarshalJSON(data []byte) error {
 			delete(rawMsg, key)
 		case "status":
 			err = unpopulate(val, "Status", &c.Status)
+			delete(rawMsg, key)
+		}
+		if err != nil {
+			return fmt.Errorf("unmarshalling type %T: %v", c, err)
+		}
+	}
+	return nil
+}
+
+// MarshalJSON implements the json.Marshaller interface for type ContinuationSettingsReference.
+func (c ContinuationSettingsReference) MarshalJSON() ([]byte, error) {
+	objectMap := make(map[string]any)
+	populateAny(objectMap, "continuationTtlInMinutes", c.ContinuationTTLInMinutes)
+	populateAny(objectMap, "customizedCheckpointKey", c.CustomizedCheckpointKey)
+	populateAny(objectMap, "idleCondition", c.IdleCondition)
+	return json.Marshal(objectMap)
+}
+
+// UnmarshalJSON implements the json.Unmarshaller interface for type ContinuationSettingsReference.
+func (c *ContinuationSettingsReference) UnmarshalJSON(data []byte) error {
+	var rawMsg map[string]json.RawMessage
+	if err := json.Unmarshal(data, &rawMsg); err != nil {
+		return fmt.Errorf("unmarshalling type %T: %v", c, err)
+	}
+	for key, val := range rawMsg {
+		var err error
+		switch key {
+		case "continuationTtlInMinutes":
+			err = unpopulate(val, "ContinuationTTLInMinutes", &c.ContinuationTTLInMinutes)
+			delete(rawMsg, key)
+		case "customizedCheckpointKey":
+			err = unpopulate(val, "CustomizedCheckpointKey", &c.CustomizedCheckpointKey)
+			delete(rawMsg, key)
+		case "idleCondition":
+			err = unpopulate(val, "IdleCondition", &c.IdleCondition)
 			delete(rawMsg, key)
 		}
 		if err != nil {
@@ -19277,6 +19363,7 @@ func (e *ExecuteDataFlowActivity) UnmarshalJSON(data []byte) error {
 func (e ExecuteDataFlowActivityTypeProperties) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]any)
 	populate(objectMap, "compute", e.Compute)
+	populate(objectMap, "continuationSettings", e.ContinuationSettings)
 	populateAny(objectMap, "continueOnError", e.ContinueOnError)
 	populate(objectMap, "dataFlow", e.DataFlow)
 	populate(objectMap, "integrationRuntime", e.IntegrationRuntime)
@@ -19298,6 +19385,9 @@ func (e *ExecuteDataFlowActivityTypeProperties) UnmarshalJSON(data []byte) error
 		switch key {
 		case "compute":
 			err = unpopulate(val, "Compute", &e.Compute)
+			delete(rawMsg, key)
+		case "continuationSettings":
+			err = unpopulate(val, "ContinuationSettings", &e.ContinuationSettings)
 			delete(rawMsg, key)
 		case "continueOnError":
 			err = unpopulate(val, "ContinueOnError", &e.ContinueOnError)
@@ -19514,6 +19604,7 @@ func (e *ExecutePipelineActivityTypeProperties) UnmarshalJSON(data []byte) error
 func (e ExecutePowerQueryActivityTypeProperties) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]any)
 	populate(objectMap, "compute", e.Compute)
+	populate(objectMap, "continuationSettings", e.ContinuationSettings)
 	populateAny(objectMap, "continueOnError", e.ContinueOnError)
 	populate(objectMap, "dataFlow", e.DataFlow)
 	populate(objectMap, "integrationRuntime", e.IntegrationRuntime)
@@ -19537,6 +19628,9 @@ func (e *ExecutePowerQueryActivityTypeProperties) UnmarshalJSON(data []byte) err
 		switch key {
 		case "compute":
 			err = unpopulate(val, "Compute", &e.Compute)
+			delete(rawMsg, key)
+		case "continuationSettings":
+			err = unpopulate(val, "ContinuationSettings", &e.ContinuationSettings)
 			delete(rawMsg, key)
 		case "continueOnError":
 			err = unpopulate(val, "ContinueOnError", &e.ContinueOnError)
@@ -42034,6 +42128,7 @@ func (s SQLServerLinkedServiceTypeProperties) MarshalJSON() ([]byte, error) {
 	populateAny(objectMap, "connectRetryInterval", s.ConnectRetryInterval)
 	populateAny(objectMap, "connectTimeout", s.ConnectTimeout)
 	populateAny(objectMap, "connectionString", s.ConnectionString)
+	populate(objectMap, "credential", s.Credential)
 	populateAny(objectMap, "database", s.Database)
 	populateAny(objectMap, "encrypt", s.Encrypt)
 	populate(objectMap, "encryptedCredential", s.EncryptedCredential)
@@ -42086,6 +42181,9 @@ func (s *SQLServerLinkedServiceTypeProperties) UnmarshalJSON(data []byte) error 
 			delete(rawMsg, key)
 		case "connectionString":
 			err = unpopulate(val, "ConnectionString", &s.ConnectionString)
+			delete(rawMsg, key)
+		case "credential":
+			err = unpopulate(val, "Credential", &s.Credential)
 			delete(rawMsg, key)
 		case "database":
 			err = unpopulate(val, "Database", &s.Database)
@@ -49774,6 +49872,7 @@ func (s SnowflakeExportCopyCommand) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]any)
 	populate(objectMap, "additionalCopyOptions", s.AdditionalCopyOptions)
 	populate(objectMap, "additionalFormatOptions", s.AdditionalFormatOptions)
+	populateAny(objectMap, "storageIntegration", s.StorageIntegration)
 	objectMap["type"] = "SnowflakeExportCopyCommand"
 	if s.AdditionalProperties != nil {
 		for key, val := range s.AdditionalProperties {
@@ -49797,6 +49896,9 @@ func (s *SnowflakeExportCopyCommand) UnmarshalJSON(data []byte) error {
 			delete(rawMsg, key)
 		case "additionalFormatOptions":
 			err = unpopulate(val, "AdditionalFormatOptions", &s.AdditionalFormatOptions)
+			delete(rawMsg, key)
+		case "storageIntegration":
+			err = unpopulate(val, "StorageIntegration", &s.StorageIntegration)
 			delete(rawMsg, key)
 		case "type":
 			err = unpopulate(val, "Type", &s.Type)
@@ -49824,6 +49926,7 @@ func (s SnowflakeImportCopyCommand) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]any)
 	populate(objectMap, "additionalCopyOptions", s.AdditionalCopyOptions)
 	populate(objectMap, "additionalFormatOptions", s.AdditionalFormatOptions)
+	populateAny(objectMap, "storageIntegration", s.StorageIntegration)
 	objectMap["type"] = "SnowflakeImportCopyCommand"
 	if s.AdditionalProperties != nil {
 		for key, val := range s.AdditionalProperties {
@@ -49847,6 +49950,9 @@ func (s *SnowflakeImportCopyCommand) UnmarshalJSON(data []byte) error {
 			delete(rawMsg, key)
 		case "additionalFormatOptions":
 			err = unpopulate(val, "AdditionalFormatOptions", &s.AdditionalFormatOptions)
+			delete(rawMsg, key)
+		case "storageIntegration":
+			err = unpopulate(val, "StorageIntegration", &s.StorageIntegration)
 			delete(rawMsg, key)
 		case "type":
 			err = unpopulate(val, "Type", &s.Type)